LayoutTests/platform/ios/media/video-source-type-params-expected.txt

EVENT(loadstart)
EXPECTED (stripExtension(relativeURL(video.currentSrc)) == 'content/test'), OBSERVED 'content/error2' FAIL
END OF TEST
